Refactored All Governors  (#61)
Refactored All Governors
