+++ integrationTest/governorTransferToken.ts
@@ -1,16 +1,14 @@
 import { BigNumber } from "bignumber.js";
 import * as fs from "fs";
 import {
-  AccountId,
+  ContractId,
   ContractExecuteTransaction,
   ContractFunctionParameters,
-  Hbar,
   TokenId,
 } from "@hashgraph/sdk";
 
 import ClientManagement from "./utils/utils";
 import { ContractService } from "../deployment/service/ContractService";
-import { ethers } from "ethers";
 import GovernorMethods from "./GovernorMethods";
 
 const clientManagement = new ClientManagement();
@@ -19,151 +17,62 @@ const contractService = new ContractService();
 const governor = new GovernorMethods();
 
 let client = clientManagement.createOperatorClient();
-const { adminId } = clientManagement.getAdmin();
 const { id, key } = clientManagement.getOperator();
 
 const { treasureId, treasureKey } = clientManagement.getTreasure();
 
-const htsServiceAddress = contractService.getContract(
-  contractService.baseContractName
-).address;
-
-const contractId = contractService.getContractWithProxy(
-  contractService.governorTTContractName
-).transparentProxyId!;
+const contractId = contractService.getContractWithProxy(contractService.governorTTContractName).transparentProxyId!;
 const transferTokenId = TokenId.fromString("0.0.48504379");
 
-const readFileContent = (filePath: string) => {
-  const rawdata: any = fs.readFileSync(filePath);
-  return JSON.parse(rawdata);
-};
-
-const initialize = async (tokenId: TokenId) => {
-  console.log(`\nInitialize contract with token  `);
-  const votingDelay = 0;
-  const votingPeriod = 12;
-
-  let contractFunctionParameters = new ContractFunctionParameters()
-    .addAddress(tokenId.toSolidityAddress()) // token that define the voting weight, to vote user should have % of this token.
-    .addAddress(id.toSolidityAddress()) // from
-    .addAddress(treasureId.toSolidityAddress()) // to
+async function propose(
+  description: string,
+  contractId: string | ContractId
+) {
+  console.log(`\nCreating proposal `);
+  const contractFunctionParameters = new ContractFunctionParameters()
+    .addString(description)
+    .addAddress(id.toSolidityAddress())// from
+    .addAddress(treasureId.toSolidityAddress())// to
     .addAddress(transferTokenId.toSolidityAddress()) // tokenToTransfer
     .addInt256(new BigNumber(100000000)) // amountToTransfer
-    .addUint256(votingDelay)
-    .addUint256(votingPeriod)
-    .addAddress(htsServiceAddress);
 
   const tx = await new ContractExecuteTransaction()
     .setContractId(contractId)
-    .setFunction("initialize", contractFunctionParameters)
-    .setGas(900000)
-    .execute(client);
-
-  const receipt = await tx.getReceipt(client);
-
-  console.log(`Initialize contract with token done with status - ${receipt}`);
-};
-
-const execute = async (
-  targets: Array<string>,
-  ethFees: Array<number>,
-  calls: Array<Uint8Array>,
-  description: string
-) => {
-  console.log(`\nExecuting  proposal - `);
-
-  const contractFunctionParameters = new ContractFunctionParameters()
-    .addAddressArray(targets)
-    .addUint256Array(ethFees)
-    .addBytesArray(calls)
-    .addString(description);
-
-  const contractAllotTx = await new ContractExecuteTransaction()
-    .setContractId(contractId)
-    .setFunction("executePublic", contractFunctionParameters)
-    .setPayableAmount(new Hbar(70))
-    .setMaxTransactionFee(new Hbar(70))
-    .setGas(900000)
+    .setFunction("createProposal", contractFunctionParameters)
+    .setGas(9000000)
     .freezeWith(client)
-    .sign(treasureKey); //Admin of token
+    .sign(treasureKey);
 
-  const executedTx = await contractAllotTx.execute(client);
+  const executedTx = await tx.execute(client);
+
   const record = await executedTx.getRecord(client);
-  const contractAllotRx = await executedTx.getReceipt(client);
-  const status = contractAllotRx.status;
+  const receipt = await executedTx.getReceipt(client);
 
-  console.log(
-    `Execute tx status ${status} for proposal id ${record.contractFunctionResult?.getUint256(
-      0
-    )}`
-  );
-};
+  const status = receipt.status;
+  const proposalId = record.contractFunctionResult?.getUint256(0)!;
+  console.log(`Proposal tx status ${status} with proposal id ${proposalId}`);
 
-const associateTokenPublicCallData = async (
-  tokenId: TokenId
-): Promise<Uint8Array> => {
-  const contractJson = readFileContent(
-    "./artifacts/contracts/common/BaseHTS.sol/BaseHTS.json"
-  );
-  const contractInterface = new ethers.utils.Interface(contractJson.abi);
-
-  const receiver = adminId.toSolidityAddress();
-  const callData = contractInterface.encodeFunctionData(
-    "associateTokenPublic",
-    [receiver, tokenId.toSolidityAddress()]
-  );
-  return ethers.utils.toUtf8Bytes(callData);
-};
-
-const quorumReached = async (proposalId: BigNumber) => {
-  console.log(`\nGetting quorumReached `);
-
-  let contractFunctionParameters = new ContractFunctionParameters().addUint256(
-    proposalId
-  );
-
-  const contractTokenTx = await new ContractExecuteTransaction()
-    .setContractId(contractId)
-    .setFunction("quorumReached", contractFunctionParameters)
-    .setGas(500000)
-    .execute(client);
-
-  const receipt = await contractTokenTx.getReceipt(client);
-  const record = await contractTokenTx.getRecord(client);
-  const status = record.contractFunctionResult!.getBool(0);
-
-  console.log(
-    `quorumReached tx status ${receipt.status} with quorumReached ${status}`
-  );
+  return proposalId;
 };
 
 async function main() {
   console.log(`\nUsing governor proxy contract id ${contractId}`);
-  const tokenId = TokenId.fromString("0.0.48602743");
-  await initialize(tokenId);
+  await governor.initialize(contractId);
+  const description = "Create token proposal 5";
 
-  const targets = [htsServiceAddress];
-  const ethFees = [0];
-  const associateToken = await associateTokenPublicCallData(tokenId);
-  const calls = [associateToken];
-  const description = "Create token proposal 2";
-
-  const proposalId = await governor.propose(
-    targets,
-    ethFees,
-    calls,
+  const proposalId = await propose(
     description,
     contractId
   );
   await governor.vote(proposalId, 1, contractId); //1 is for vote.
-  await quorumReached(proposalId);
+  await governor.quorumReached(proposalId, contractId);
   await governor.voteSucceeded(proposalId, contractId);
   await governor.proposalVotes(proposalId, contractId);
   await governor.state(proposalId, contractId);
   console.log(`\nWaiting for voting period to get over.`);
   await new Promise((f) => setTimeout(f, 15 * 1000)); //Wait till waiting period is over. It's current deadline as per Governance.
   await governor.state(proposalId, contractId); //4 means succeeded
-  await execute(targets, ethFees, calls, description);
+  await governor.execute(description, contractId);
   console.log(`\nDone`);
 }
